moved token validator to middleware folder
